C functions

+++ 0x00-hello_world/6-size.c
@@ -15,11 +15,11 @@ int main(void)
 	long long alonglong;
 	float afloat;
 
-	printf("Size of a char: %ld byte(s)\n", sizeof(acharacter));
-	printf("Size of an int: %ld byte(s)\n", sizeof(aninteger));
-	printf("Size of a long int: %ld byte(s)\n", sizeof(along));
-	printf("Size of a long long int: %ld byte(s)\n", sizeof(alonglong));
-	printf("Size of a float: %ld byte(s)\n", sizeof(afloat));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(acharacter));
+	printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(aninteger));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(along));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(alonglong));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(afloat));
 
 
 	return (0);
